Move SoundcloudExtractorHelper methods to SoundcloudParsingHelper
