+++ src/main/java/org/schabi/newpipe/extractor/services/soundcloud/SoundcloudParsingHelper.java
@@ -21,9 +21,14 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
+import javax.annotation.Nonnull;
+
 public class SoundcloudParsingHelper {
     private static String clientId;
 
+    private static final String HTTP = "http://";
+    private static final String HTTPS = "https://";
+    
     private SoundcloudParsingHelper() {
     }
 
@@ -157,5 +162,28 @@ public class SoundcloudParsingHelper {
 
     public static String getStreamsFromApi(StreamInfoItemCollector collector, String apiUrl) throws ReCaptchaException, ParsingException, IOException {
         return getStreamsFromApi(collector, apiUrl, false);
+    }
+
+    private static String replaceHttpWithHttps(final String url) {
+        if(!url.isEmpty() && url.startsWith(HTTP)) {
+            return HTTPS + url.substring(HTTP.length());
+        }
+        return url;
+    }
+
+    @Nonnull
+    static String getUploaderUrl(JsonObject object) {
+        String url = object.getObject("user").getString("permalink_url", "");
+        return replaceHttpWithHttps(url);
+    }
+
+    @Nonnull
+    static String getAvatarUrl(JsonObject object) {
+        String url = object.getObject("user", new JsonObject()).getString("avatar_url", "");
+        return replaceHttpWithHttps(url);
+    }
+
+    public static String getUploaderName(JsonObject object) {
+        return object.getObject("user").getString("username", "");
     }
 }
